+++ validation/validation.py
@@ -10,11 +10,13 @@ MODEL = "ro_core_news_sm"
 
 def validate_dataset(dataset_type: str, mock_data: bool = False):
     """
-    Runs a validation loop over a specified dataset (e.g., ``validation``, ``test``, or ``train``),
-    applying a spaCy model to all temporal expressions and writing the results to file.
+    Runs a validation loop over a specified dataset (e.g., ``validation``, ``test``,
+    or ``train``), applying a spaCy model to all temporal expressions and writing the
+    results to file.
 
-    The function loads RONEC entries, extracts their associated timespans (temporal expressions),
-    processes each using the NLP model, and writes the annotated output using ``OutputFile``.
+    The function loads RONEC entries, extracts their associated timespans (temporal
+    expressions), processes each using the NLP model, and writes the annotated output
+    using ``OutputFile``.
 
     Args:
         dataset_type (str): The name of the dataset split to validate. Expected values:
@@ -43,8 +45,8 @@ def validate_dataset(dataset_type: str, mock_data: bool = False):
     ronec = ronec_example if mock_data else load_dataset("ronec")
     OutputFile.write_header(dataset_type)
     dataset = ronec[dataset_type]
-    counter = 0
-    counter_timespan = 0
+    total_rows = 0
+    total_timespans = 0
 
     print(f"{dataset_type}: no. of entries = {len(dataset)}")
 
@@ -53,15 +55,17 @@ def validate_dataset(dataset_type: str, mock_data: bool = False):
 
         if ronec_entry.timespans:
             print(f"i = {i}")
-            counter += 1
+            total_rows += 1
 
             for timespan in ronec_entry.timespans:
                 doc = nlp(timespan.text)
-                OutputFile.write_entities_entries(dataset_type, ronec_entry, timespan, doc)
-                counter_timespan += 1
+                OutputFile.write_entities_entries(
+                    dataset_type, ronec_entry, timespan, doc
+                )
+                total_timespans += 1
 
-    print(f"{dataset_type}: no. of date and periods entries = {counter}")
-    print(f"{dataset_type}: TOTAL no. of date and periods = {counter_timespan}")
+    print(f"{dataset_type}: no. of date and periods entries = {total_rows}")
+    print(f"{dataset_type}: TOTAL no. of date and periods = {total_timespans}")
 
 
 if __name__ == "__main__":