+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 ---
 
+### 2.0.2
+#### Changed
+- Improved the log of temporal expressions that are not processed by the TeN Framework.
+
+### 2.0.1
+#### Changed
+- Updated the text of `Ronec` entities by adding space between the corresponding tokens
+- Handled ambiguous or loosely defined expressions, such as "martie -iunie 2013"
+
 ### 2.0.0.post2
 #### Added
 - icon.png.